added fuctionality to seed data, routes, controllers etc.

+++ controllers/userController.js
@@ -1,9 +1,11 @@
-const { User } = require('../models');
+const { User, Thought } = require('../models');
 
 module.exports = {
-    getUsers: async (res, req) => {
+    getUsers: async (req, res) => {
         try {
-            const usersData = await User.find({});
+            const usersData = await User.find({}).populate('thoughts')
+            .select('-__v')
+            .populate({path : 'friends', select: 'username'});;
             res.json(usersData);
         } catch (err) {
             console.log(err);
@@ -11,9 +13,12 @@ module.exports = {
         }
     },
 
-    getUserbyId: async (res, req) => {
+    getUserbyId: async (req, res) => {
         try {
-            const userData = await User.findOne({ _id: req.params.id });
+            const userData = await User.findOne({ _id: req.params.id })
+            .populate('thoughts')
+            .select('-__v')
+            .populate({path : 'friends', select: 'username'});
             res.json(userData);
         } catch (err) {
             console.log(err);
@@ -21,7 +26,7 @@ module.exports = {
         }
     },
 
-    createUser: async (res, req) => {
+    createUser: async (req, res) => {
         try {
             const userData = await User.create(req.body);
             res.json(userData);
@@ -31,7 +36,7 @@ module.exports = {
         }
     },
 
-    updateUser: async (res, req) => {
+    updateUser: async (req, res) => {
         try {
             const userData = await User.findOneAndUpdate(
                 { _id: req.params.id },
@@ -49,14 +54,14 @@ module.exports = {
         }
     },
 
-    deleteUser: async (res, req) => {
+    deleteUser: async (req, res) => {
         try {
-            const userData = await User.findOneAndDelete({ _id: req.params.id });
+            const userData = await User.findOneAndDelete({ _id: req.params.id || req.body.userId});
             if (!userData) {
                 res.status(404).json({ message: 'No user found with this id!' });
                 return;
             }
-            res.json(userData);
+            res.json("User deleted");
         } catch (err) {
             console.log(err);
             res.status(500).json(err);
@@ -65,11 +70,11 @@ module.exports = {
     },
 
 
-    addFriend: async (res, req) => {
+    addFriend: async (req, res) => {
         try {
             const userData = await User.findOneAndUpdate(
-                { _id: req.params.id },
-                { $push: { friends: req.params.friendId } },
+                { _id: req.params.userId },
+                { $push: { friends: req.params.friendId   } },
                 { new: true }
             );
             if (!userData) {
@@ -83,19 +88,18 @@ module.exports = {
         }
     },
 
-    deleteFriend: async (res, req) => {
+    deleteFriend: async (req, res) => {
         try {
-            const userData = await User.findOneAndUpdate({
-                _id: req.params.id
-            },
-                { $pull: { friends: req.params.friendId } },
+            const userData = await User.findOneAndUpdate(
+                {_id: req.params.userId},
+                { $pull: { friends:  req.params.friendId} },
                 { new: true }
             );
             if (!userData) {
                 res.status(404).json({ message: 'No user found with this id!' });
                 return;
             }
-            res.json(userData);
+            res.json("friend deleted /n"+ userData);
         } catch (err) {
             console.log(err);
             res.status(500).json(err);
